refactor(kitchenpos): 메뉴(Menu) 도메인 계층으로 로직 이전
- 메뉴 가격 비교는 엔티티 저장시 검증

+++ src/main/java/kitchenpos/dto/MenuResponse.java
@@ -10,11 +10,11 @@ import java.util.stream.Collectors;
 public class MenuResponse {
     private Long id;
     private String name;
-    private BigDecimal price;
+    private long price;
     private MenuGroupResponse menuGroup;
     private List<MenuProductResponse> menuProducts;
 
-    public MenuResponse(Long id, String name, BigDecimal price, MenuGroupResponse menuGroup, List<MenuProductResponse> menuProducts) {
+    public MenuResponse(Long id, String name, long price, MenuGroupResponse menuGroup, List<MenuProductResponse> menuProducts) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -41,7 +41,7 @@ public class MenuResponse {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public long getPrice() {
         return price;
     }
